requirements.txt
autopep8==1.6.0
pathspec==0.9.0
pycodestyle==2.8.0
PyYAML==6.0
toml==0.10.2
yamllint==1.26.3
